configure default location for time parsing
Interacting with servers that utilize the ISO-8601 Local Time variants in
request and response payloads can configure the request marshaling using
NormalizeTimeForMarshal setting, but don't have a ready method for setting the
local time zone when parsing the response fields.

Keep UTC as the default location, to maintain the previous functionality.